requirements.txt
streamlit==1.28.1
pandas==2.1.3
python-docx==0.8.11
PyPDF2==3.0.1
openpyxl==3.1.2
plotly==5.17.0
streamlit-aggrid==0.3.4
watchdog==3.0.0
regex==2023.10.3
jsonschema==4.20.0
pydantic==2.5.0
markdown==3.4.4